Refactor FXIOS-6534 [v116] Clean up debug options in AppSettingsOptions (#14769)
* Clean up debug options in AppSettingsOptions

- Put each class in it's own file
- Debug option now appears in their own section when triggered, so no need to have "Debug: " in the title
- Added a AppSettingsDelegate since the number of clicks on the version to trigger the debug options was a private var in top of the file. Now it's in the VC instead
- Reordered the debug options in alphabetical order, apart for Experiments which is now at the top. Open 50 mozilla tabs is at the bottom.
- Remove SlowDataBase settings option for now

* Fix and add version setting test + create Settings test folder

* Remove comment
