+++ Client/Frontend/Settings/Main/Debug/AppReviewPromptSetting.swift
@@ -1,0 +1,17 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+import Shared
+
+class AppReviewPromptSetting: HiddenSetting {
+    override var title: NSAttributedString? {
+        return NSAttributedString(string: "Toggle App Review (needs tab switch)", attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
+    }
+
+    override func onClick(_ navigationController: UINavigationController?) {
+        UserDefaults.standard.set(true, forKey: PrefsKeys.ForceShowAppReviewPromptOverride)
+        updateCell(navigationController)
+    }
+}
